requirements.txt
gradio
pandas
numpy
matplotlib
Pillow
PyPDF2
scikit-learn